src/main/java/LeetCode/Leet103.java
package LeetCode;

import java.util.*;

/**
 * Binary Tree Zigzag Level Order Traversal
 * https://leetcode.com/problems/binary-tree-zigzag-level-order-traversal/
 * created: 2021-07-30
 */
public class Leet103 {
    public List<List<Integer>> zigzagLevelOrder(TreeNode root) {
        Queue<TreeNode> queue = new LinkedList<>();
        List<List<Integer>> answer = new ArrayList<>();
        queue.offer(root);

        if (root == null) return answer;
        boolean reverse = false;

        while (!queue.isEmpty()) {
            int size = queue.size();
            List<Integer> list = new ArrayList<>();

            for (int i = 0; i < size; i++) {
                TreeNode current = queue.poll();
                if (current != null) {
                    list.add(current.val);
                    if (current.left != null) queue.offer(current.left);
                    if (current.right != null) queue.offer(current.right);
                }
            }
            if (reverse) Collections.reverse(list);
            answer.add(list);
            reverse = !reverse;
        }
        System.out.println(answer);
        return answer;
    }
}
